Impl nested record in iframe
In this implementation, we add a root id to the serialized node
when its target document is not the top scope document object.
Currently we do not rely on the root id in rebuilding, but has an
assertion to make sure we are handling this right.
BTW, iframe contentDocument will only be available when it was
append to the DOM tree. Since we are building nodes recursively,
we put iframe stuffs into a callback array which will execute after
the whole rebuilding.